Eliminate destination global variable
We now cart a pointer around everywhere. This opens the possibility of
instantiating multiple destination objects in the future, though that's
unlikely. It's more likely that we'll make the contents of the struct
destination opaque and require all access to go through a method that
takes one of these new instance pointers..