Rejig walker to take a generic closure
We passed a squares repository pointer beginning in commit f182d97 as
it was convenient at that time; we now implement a more generic solution
(we should be able to walk any repository without having a squares
repository open already).

We also take this opportunity to have the import walker use an
import_job struct as its closure.